requirements.txt
customtkinter>=5.2.0
PyPDF2>=3.0.0
pillow>=10.0.0
python-dateutil>=2.8.0
pytest>=7.0.0
